+++ test/libsolidity/semanticTests/immutable/read_in_ctor.sol
@@ -1,0 +1,17 @@
+contract A {
+	uint8 immutable a;
+	uint8 x;
+
+	constructor() {
+		a = 3;
+		x = a;
+	}
+
+	function readX() public view returns (uint8) {
+		return x;
+	}
+}
+// ====
+// compileViaYul: also
+// ----
+// readX() -> 3