+++ test/libsolidity/syntaxTests/immutable/ctor_modifier_reading.sol
@@ -11,4 +11,4 @@ contract C {
     function f(uint a) internal pure {}
 }
 // ----
-// TypeError 7733: (119-120): Immutable variables cannot be read during contract creation time, which means they cannot be read in the constructor or any function or modifier called from it.
+// TypeError 7733: (119-120): Immutable variables cannot be read before they are initialized.