+++ test/libsolidity/syntaxTests/immutable/increment.sol
@@ -5,4 +5,4 @@ contract C {
     }
 }
 // ----
-// TypeError 7733: (63-64): Immutable variables cannot be read during contract creation time, which means they cannot be read in the constructor or any function or modifier called from it.
+// TypeError 3969: (63-64): Immutable variables must be initialized using an assignment.